INVT-11610: Update snakeyaml to version 2.2 (#74)
This one is a derived dependency of spring-boot, and we are already in the latest spring boot version. This is a false positive as stated here https://github.com/spring-projects/spring-boot/issues/33457 and since we are not doing any manual yaml parsing. However, we can update to v2.2 to completely remove the snakeyaml 1.33 from our jars.